Removoe dynamic features
