+++ app/build.gradle
@@ -73,8 +73,6 @@ android {
   packagingOptions {
     exclude 'META-INF/licenses/ASM'
   }
-
-  dynamicFeatures = [":feature:about", ":feature:settings"]
 }
 
 dependencies {
@@ -85,12 +83,11 @@ dependencies {
   implementation project(':navigation')
   implementation project(':lib:navigation-deeplink')
   implementation project(':lib:identity')
-  implementation project(':lib:dynamic-features')
   implementation project(':lib:network-status')
   implementation project(':feature:push')
   implementation project(':feature:users')
-  implementation project(':feature:settings_entrance')
-  implementation project(':feature:about_entrance')
+  implementation project(':feature:settings')
+  implementation project(':feature:about')
   implementation project(':feature:shortcuts')
 
   implementation 'androidx.appcompat:appcompat:1.1.0'
@@ -118,7 +115,6 @@ dependencies {
 
   implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
   implementation 'com.jakewharton.threetenabp:threetenabp:1.2.4'
-  implementation 'com.google.android.play:core:1.6.1'
   implementation 'com.squareup.okhttp3:logging-interceptor:4.6.0'
 
   testImplementation 'com.google.code.gson:gson:2.8.6'
