--- app/src/main/java/com/jraska/github/client/HasDynamicFeaturesComponent.kt
@@ -1,9 +1,0 @@
-package com.jraska.github.client
-
-import androidx.fragment.app.FragmentActivity
-
-interface HasDynamicFeaturesComponent {
-  fun dynamicFeaturesComponent(): DynamicFeaturesComponent
-}
-
-fun FragmentActivity.dynamicFeaturesComponent() = (application as HasDynamicFeaturesComponent).dynamicFeaturesComponent()
