--- feature/about/src/main/java/com/jraska/github/client/about/AboutFeature.kt
@@ -1,12 +1,0 @@
-package com.jraska.github.client.about
-
-import androidx.annotation.Keep
-import com.jraska.github.client.dynamicbase.DynamicFeature
-import timber.log.Timber
-
-@Keep // Used by reflection
-class AboutFeature : DynamicFeature {
-  override fun onFeatureCreate() {
-    Timber.i("about feature loaded")
-  }
-}
