--- feature/about_entrance/.gitignore
@@ -1,1 +1,0 @@
-/build
