--- feature/about_entrance/build.gradle
@@ -1,27 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
-
-android {
-  compileSdkVersion 29
-  defaultConfig {
-    minSdkVersion 21
-  }
-
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_8
-    targetCompatibility JavaVersion.VERSION_1_8
-  }
-}
-
-dependencies {
-  implementation project(':core')
-  implementation project(':core-android')
-
-  implementation 'com.google.dagger:dagger:2.27'
-  kapt 'com.google.dagger:dagger-compiler:2.27'
-  implementation project(':lib:dynamic-features')
-
-  testImplementation 'junit:junit:4.13'
-  testImplementation 'org.assertj:assertj-core:3.15.0'
-}
