--- feature/about_entrance/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.jraska.github.client.about.entrance" />
