--- feature/about_entrance/src/main/java/com/jraska/github/client/about/entrance/internal/ReflectiveAboutFeatureProvider.kt
@@ -1,10 +1,0 @@
-package com.jraska.github.client.about.entrance.internal
-
-import com.jraska.github.client.dynamicbase.DynamicFeature
-import javax.inject.Provider
-
-class ReflectiveAboutFeatureProvider : Provider<DynamicFeature> {
-  override fun get(): DynamicFeature {
-    return Class.forName("com.jraska.github.client.about.AboutFeature").newInstance() as DynamicFeature
-  }
-}
