--- feature/about_entrance/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-  <string name="title_dynamic_feature_about" translatable="false">about</string>
-</resources>
