+++ feature/settings/build.gradle
@@ -1,4 +1,4 @@
-apply plugin: 'com.android.dynamic-feature'
+apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-android-extensions'
@@ -18,13 +18,6 @@ android {
 dependencies {
   implementation project(':core')
   implementation project(':core-android')
-  implementation project(':lib:dynamic-features')
-  implementation project(':app')
-
-  implementation project(':lib:dynamic-features')
-
-  implementation 'com.google.android.play:core:1.6.1'
-  implementation 'com.google.android.gms:play-services-basement:17.1.1'
 
   implementation 'com.airbnb.android:epoxy:3.9.0'
   implementation 'com.jraska:console:1.1.0'
@@ -35,7 +28,4 @@ dependencies {
 
   testImplementation 'junit:junit:4.13'
   testImplementation 'org.assertj:assertj-core:3.15.0'
-
-  androidTestImplementation 'androidx.appcompat:appcompat:1.1.0'
-  androidTestImplementation 'com.google.android.gms:play-services-basement:17.1.1'
 }
