--- feature/settings_entrance/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.jraska.github.client.settings.entrance" />
