--- lib/dynamic-features/build.gradle
@@ -1,28 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
-
-android {
-  compileSdkVersion 29
-  defaultConfig {
-    minSdkVersion 21
-  }
-
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_8
-    targetCompatibility JavaVersion.VERSION_1_8
-  }
-}
-
-dependencies {
-  implementation project(':core')
-  implementation project(':core-android')
-
-  kapt 'com.google.dagger:dagger-compiler:2.27'
-  implementation 'com.google.dagger:dagger:2.27'
-  implementation 'com.google.android.play:core:1.6.1'
-
-  testImplementation 'junit:junit:4.13'
-  testImplementation project(':core-testing')
-  testImplementation 'org.assertj:assertj-core:3.15.0'
-}
