--- lib/dynamic-features/src/main/java/com/jraska/github/client/dynamicbase/DynamicFeatureInstaller.kt
@@ -1,7 +1,0 @@
-package com.jraska.github.client.dynamicbase
-
-import io.reactivex.Completable
-
-interface DynamicFeatureInstaller {
-  fun ensureInstalled(featureName: String): Completable
-}
