--- lib/dynamic-features/src/main/java/com/jraska/github/client/dynamicbase/DynamicFeatureSpec.kt
@@ -1,8 +1,0 @@
-package com.jraska.github.client.dynamicbase
-
-import javax.inject.Provider
-
-class DynamicFeatureSpec(
-  val featureName: String,
-  val featureProvider: Provider<DynamicFeature>
-)
