--- lib/dynamic-features/src/main/java/com/jraska/github/client/dynamicbase/internal/SplitInstallManager.kt
@@ -1,14 +1,0 @@
-package com.jraska.github.client.dynamicbase.internal
-
-import com.google.android.play.core.splitinstall.SplitInstallManager
-import com.google.android.play.core.splitinstall.SplitInstallRequest
-
-internal fun SplitInstallManager.beginInstallation(featureName: String) {
-  val request = SplitInstallRequest.newBuilder()
-    .addModule(featureName)
-    .build()
-
-  startInstall(request)
-}
-
-internal fun SplitInstallManager.alreadyInstalled(featureName: String) = this.installedModules.contains(featureName)
