--- lib/dynamic-features/src/main/res/layout/activity_feature_install.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  android:padding="16dp"
-  android:minWidth="256dp"
-  android:minHeight="96dp"
-  android:orientation="vertical"
-  >
-
-  <TextView
-    android:id="@+id/feature_install_progress_label"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:gravity="center_horizontal"
-    tools:text="Downloading about feature"
-    />
-
-  <ProgressBar
-    android:id="@+id/feature_install_progress_bar"
-    style="@style/Widget.AppCompat.ProgressBar.Horizontal"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:indeterminate="true"
-    />
-
-</LinearLayout>
