--- lib/dynamic-features/src/main/res/values/strings.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-  <string name="feature_downloading">Downloading %s</string>
-  <string name="feature_pending">Pending install for %s</string>
-  <string name="feature_installing">Installing %s</string>
-  <string name="feature_module_name">Error installing %s</string>
-</resources>
